+++ src/app/gear-grid/gear-grid.actions.ts
@@ -1,5 +1,5 @@
 import {createAction, props} from '@ngrx/store';
-import {Item} from '../../gear-grid-logic/item';
+import {Item} from '../gear-grid-logic/item';
 
 export const loadGearGrids = createAction(
     '[GearGrid] Load GearGrids'
