+++ src/app/gear-grid/gear-grid.module.ts
@@ -1,8 +1,8 @@
 import {NgModule} from '@angular/core';
 import {StoreModule} from '@ngrx/store';
-import * as fromGearGrid from './reducers/gear-grid.reducer';
+import * as fromGearGrid from './gear-grid.reducer';
 import {EffectsModule} from '@ngrx/effects';
-import {GearGridEffects} from './effects/gear-grid.effects';
+import {GearGridEffects} from './gear-grid.effects';
 import {GearGridContainerComponent} from './containers/gear-grid-container/gear-grid-container.component';
 import {GearItemComponent} from './components/gear-item/gear-item.component';
 import {GearGridComponent} from './components/gear-grid/gear-grid.component';
