+++ src/app/gear-grid/gear-grid.selectors.ts
@@ -1,6 +1,6 @@
 import {createSelector} from '@ngrx/store';
 import {GearGridFeatureState, selectGearGridFeature} from './gear-grid.reducer';
-import {Item} from '../../gear-grid-logic/item';
+import {Item} from '../gear-grid-logic/item';
 
 export const getGearGridState = createSelector(
   selectGearGridFeature,
