+++ src/app/reducers/index.ts
@@ -1,6 +1,6 @@
 import {ActionReducerMap, MetaReducer} from '@ngrx/store';
 import {environment} from '../../environments/environment';
-import * as fromGearGrid from '../gear-grid/reducers/gear-grid.reducer';
+import * as fromGearGrid from '../gear-grid/gear-grid.reducer';
 import * as fromItems from '../items/reducers/item.reducer';
 
 export interface AppState {
